# perf: avoid loadFullConfig when creating block hoist plugin